added tables to side panel; added live calculation of advices

+++ src/main/java/org/taurus/aya/server/entity/Event.java
@@ -193,12 +193,12 @@ public class Event {
     }
 
 
-    public Integer getExecutor() {
-        return task.getExecutor().intValue();
+    public Long getExecutor() {
+        return task.getExecutor();
     }
 
-    public void setExecutor(Integer executor) {
-        task.setExecutor(executor.longValue());
+    public void setExecutor(Long executor) {
+        task.setExecutor(executor);
     }
